+++ src/openassetio-python/tests/package/pluginSystem/resources/pathC/modulePlugin.py
@@ -25,4 +25,4 @@ class ModulePlugin(PythonPluginSystemManagerPlugin):
 
 
 # pylint: disable=invalid-name
-plugin = ModulePlugin
+openassetioPlugin = ModulePlugin